+++ src/ray.rs
@@ -4,11 +4,11 @@ use vec3::{Vec3};
 pub struct Ray {
     pub origin: Vec3,
     pub direction: Vec3,
-    pub time: f64
+    pub time: f32
 }
 
 impl Ray {
-    pub fn point_at_parameter(&self, t: f64) -> Vec3 {
+    pub fn point_at_parameter(&self, t: f32) -> Vec3 {
         self.origin + self.direction * t
     }
 }
